Выполняет рефакторинг модуля с утилитами
