+++ src/view/task-edit.js
@@ -1,5 +1,5 @@
 import {COLORS} from "../const.js";
-import {isTaskExpired, isTaskRepeating, humanizeTaskDueDate} from "../utils.js";
+import {isTaskExpired, isTaskRepeating, humanizeTaskDueDate} from "../utils/task.js";
 import AbstractView from "./abstract.js";
 
 const BLANK_TASK = {
